Moved /games to root, and collaps finished games.
